+++ Source/Krypton Toolkit/Krypton Toolkit Suite Extended.sln
@@ -219,6 +219,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Arrows", "Arrows", "{0451E8
 		..\..\Assets\PSD\Arrows\Template.psd = ..\..\Assets\PSD\Arrows\Template.psd
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Krypton.Toolkit.Suite.Extended.Scroll.Bars", "Main\Krypton.Toolkit.Suite.Extended.Scroll.Bars\Krypton.Toolkit.Suite.Extended.Scroll.Bars.csproj", "{69D1AC83-0BA2-429A-B1EA-7E0E65F37451}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -437,6 +439,10 @@ Global
 		{A4A20AC1-DC76-48AC-9E91-982581AF7D24}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A4A20AC1-DC76-48AC-9E91-982581AF7D24}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A4A20AC1-DC76-48AC-9E91-982581AF7D24}.Release|Any CPU.Build.0 = Release|Any CPU
+		{69D1AC83-0BA2-429A-B1EA-7E0E65F37451}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{69D1AC83-0BA2-429A-B1EA-7E0E65F37451}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{69D1AC83-0BA2-429A-B1EA-7E0E65F37451}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{69D1AC83-0BA2-429A-B1EA-7E0E65F37451}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -510,6 +516,7 @@ Global
 		{378BEE59-0131-4497-80B8-69B3124325D7} = {734B2F67-D055-44B2-818A-FE569C479937}
 		{69F98672-EBB9-4B12-BD04-45AEF3C37FC6} = {15346861-5444-49D7-9686-3835EE271259}
 		{0451E883-C718-4679-962F-E2825B1513B2} = {247BBF4E-22B0-49CC-981B-0420AD7FF4E8}
+		{69D1AC83-0BA2-429A-B1EA-7E0E65F37451} = {252CCB85-EC69-4B0F-9ACB-4C37AE3F54B5}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {A7CE7732-0982-4703-8EFD-7715D569BAB5}
