+++ Source/Krypton Toolkit/Main/Krypton.Toolkit.Suite.Extended.Controls/Krypton.Toolkit.Suite.Extended.Controls.csproj
@@ -27,20 +27,6 @@
     <PackageReference Include="Krypton.Toolkit.Canary" Version="6.0.2103.6" />
     <!--<PackageReference Include="Krypton.Toolkit" Version="5.550.2103" />-->
   </ItemGroup>
-  <ItemGroup>
-    <Compile Update="Components\Scroll Bars\Skins\HScrollSkin.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Update="Components\Scroll Bars\Skins\VScrollSkin.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Update="Controls Toolkit\KryptonComboBoxEnhanced.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Update="Controls Toolkit\KryptonScrollBar.cs">
-      <SubType>Component</SubType>
-    </Compile>
-  </ItemGroup>
 
   <!--Binary/Documentation Output-->
   <PropertyGroup>
