+++ Source/Krypton Toolkit/Main/Krypton.Toolkit.Suite.Extended.Scroll.Bars/Classes/ScrollBarKryptonRenderer.cs
@@ -3,14 +3,14 @@ using System.Drawing;
 using System.Drawing.Drawing2D;
 using System.Drawing.Imaging;
 
-namespace Krypton.Toolkit.Suite.Extended.Controls
+namespace Krypton.Toolkit.Suite.Extended.Scroll.Bars
 {
     /// <summary>
     /// The scrollbar renderer class.
     /// </summary>
-    internal static class ScrollBarExtendedRenderer
+    internal static class ScrollBarKryptonRenderer
     {
-        #region Fields
+        #region fields
 
         /// <summary>
         /// The colors of the thumb in the 3 states.
@@ -54,9 +54,9 @@ namespace Krypton.Toolkit.Suite.Extended.Controls
         #region constructor
 
         /// <summary>
-        /// Initializes static members of the <see cref="ScrollBarExtendedRenderer"/> class.
+        /// Initializes static members of the <see cref="ScrollBarKryptonRenderer"/> class.
         /// </summary>
-        static ScrollBarExtendedRenderer()
+        static ScrollBarKryptonRenderer()
         {
             InitColours();
         }
@@ -75,34 +75,34 @@ namespace Krypton.Toolkit.Suite.Extended.Controls
 
             //Init Colors
             // hot state
-            thumbColours[0, 0] = Color.FromArgb(96, 111, 148); // border color
-            thumbColours[0, 1] = Color.FromArgb(232, 233, 233); // left/top start color
-            thumbColours[0, 2] = Color.FromArgb(230, 233, 241); // left/top end color
-            thumbColours[0, 3] = Color.FromArgb(233, 237, 242); // right/bottom line color
-            thumbColours[0, 4] = Color.FromArgb(209, 218, 228); // right/bottom start color
-            thumbColours[0, 5] = Color.FromArgb(218, 227, 235); // right/bottom end color
-            thumbColours[0, 6] = Color.FromArgb(190, 202, 219); // right/bottom middle color
-            thumbColours[0, 7] = Color.FromArgb(96, 11, 148); // left/top line color
+            thumbColours[0, 0] = _palette.GetBorderColor1(PaletteBorderStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(96, 111, 148); // border color
+            thumbColours[0, 1] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(232, 233, 233); // left/top start color
+            thumbColours[0, 2] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(230, 233, 241); // left/top end color
+            thumbColours[0, 3] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(233, 237, 242); // right/bottom line color
+            thumbColours[0, 4] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(209, 218, 228); // right/bottom start color
+            thumbColours[0, 5] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(218, 227, 235); // right/bottom end color
+            thumbColours[0, 6] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(190, 202, 219); // right/bottom middle color
+            thumbColours[0, 7] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(96, 11, 148); // left/top line color
 
             // over state
-            thumbColours[1, 0] = Color.FromArgb(60, 110, 176);
-            thumbColours[1, 1] = Color.FromArgb(187, 204, 228);
-            thumbColours[1, 2] = Color.FromArgb(205, 227, 254);
-            thumbColours[1, 3] = Color.FromArgb(252, 253, 255);
-            thumbColours[1, 4] = Color.FromArgb(170, 207, 247);
-            thumbColours[1, 5] = Color.FromArgb(219, 232, 251);
-            thumbColours[1, 6] = Color.FromArgb(190, 202, 219);
-            thumbColours[1, 7] = Color.FromArgb(233, 233, 235);
+            thumbColours[1, 0] = _palette.GetBorderColor1(PaletteBorderStyle.ButtonCluster, PaletteState.Normal);//Color.FromArgb(60, 110, 176);
+            thumbColours[1, 1] = _palette.GetBackColor2(PaletteBackStyle.ButtonCluster, PaletteState.Normal); //Color.FromArgb(187, 204, 228);
+            thumbColours[1, 2] = _palette.GetBackColor1(PaletteBackStyle.ButtonCluster, PaletteState.Normal);  //Color.FromArgb(205, 227, 254);
+            thumbColours[1, 3] = _palette.GetBackColor2(PaletteBackStyle.ButtonCluster, PaletteState.Normal);  //Color.FromArgb(252, 253, 255);
+            thumbColours[1, 4] = _palette.GetBackColor1(PaletteBackStyle.ButtonCluster, PaletteState.Normal);  //Color.FromArgb(170, 207, 247);
+            thumbColours[1, 5] = _palette.GetBackColor2(PaletteBackStyle.ButtonAlternate, PaletteState.Normal);  //Color.FromArgb(219, 232, 251);
+            thumbColours[1, 6] = _palette.GetBackColor2(PaletteBackStyle.ButtonCluster, PaletteState.Normal);  //Color.FromArgb(190, 202, 219);
+            thumbColours[1, 7] = _palette.GetBackColor2(PaletteBackStyle.ButtonCluster, PaletteState.Normal);  //Color.FromArgb(233, 233, 235);
 
             // pressed state
-            thumbColours[2, 0] = Color.FromArgb(23, 73, 138);
-            thumbColours[2, 1] = Color.FromArgb(154, 184, 225);
-            thumbColours[2, 2] = Color.FromArgb(166, 202, 250);
-            thumbColours[2, 3] = Color.FromArgb(221, 235, 251);
-            thumbColours[2, 4] = Color.FromArgb(110, 166, 240);
-            thumbColours[2, 5] = Color.FromArgb(194, 218, 248);
-            thumbColours[2, 6] = Color.FromArgb(190, 202, 219);
-            thumbColours[2, 7] = Color.FromArgb(194, 211, 231);
+            thumbColours[2, 0] = _palette.GetBorderColor1(PaletteBorderStyle.ButtonStandalone, PaletteState.CheckedNormal);//Color.FromArgb(23, 73, 138);
+            thumbColours[2, 1] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.CheckedNormal); //Color.FromArgb(154, 184, 225);
+            thumbColours[2, 2] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.CheckedNormal); // Color.FromArgb(166, 202, 250);
+            thumbColours[2, 3] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.CheckedNormal);//Color.FromArgb(221, 235, 251);
+            thumbColours[2, 4] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.CheckedNormal); //Color.FromArgb(110, 166, 240);
+            thumbColours[2, 5] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.CheckedNormal); //Color.FromArgb(194, 218, 248);
+            thumbColours[2, 6] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.CheckedNormal); //Color.FromArgb(190, 202, 219);
+            thumbColours[2, 7] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.CheckedNormal); //Color.FromArgb(194, 211, 231);
 
             /* picture of colors and indices
              *(0,0)
@@ -126,51 +126,51 @@ namespace Krypton.Toolkit.Suite.Extended.Controls
              */
 
             // hot state
-            arrowColours[0, 0] = Color.FromArgb(223, 236, 252);
-            arrowColours[0, 1] = Color.FromArgb(207, 225, 248);
-            arrowColours[0, 2] = Color.FromArgb(245, 249, 255);
-            arrowColours[0, 3] = Color.FromArgb(237, 244, 252);
-            arrowColours[0, 4] = Color.FromArgb(244, 249, 255);
-            arrowColours[0, 5] = Color.FromArgb(244, 249, 255);
-            arrowColours[0, 6] = Color.FromArgb(251, 253, 255);
-            arrowColours[0, 7] = Color.FromArgb(251, 253, 255);
+            arrowColours[0, 0] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal);//Color.FromArgb(223, 236, 252);
+            arrowColours[0, 1] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(207, 225, 248);
+            arrowColours[0, 2] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(245, 249, 255);
+            arrowColours[0, 3] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(237, 244, 252);
+            arrowColours[0, 4] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(244, 249, 255);
+            arrowColours[0, 5] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(244, 249, 255);
+            arrowColours[0, 6] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(251, 253, 255);
+            arrowColours[0, 7] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(251, 253, 255);
 
             // over state
-            arrowColours[1, 0] = Color.FromArgb(205, 222, 243); //Colore bottone sul tracking
-            arrowColours[1, 1] = Color.FromArgb(186, 208, 235);
-            arrowColours[1, 2] = Color.FromArgb(238, 244, 252);
-            arrowColours[1, 3] = Color.FromArgb(229, 237, 247);
-            arrowColours[1, 4] = Color.FromArgb(223, 234, 247);
-            arrowColours[1, 5] = Color.FromArgb(241, 246, 254);
-            arrowColours[1, 6] = Color.FromArgb(243, 247, 252);
-            arrowColours[1, 7] = Color.FromArgb(250, 252, 255);
+            arrowColours[1, 0] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal);//Color.FromArgb(205, 222, 243); //Colore bottone sul tracking
+            arrowColours[1, 1] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(186, 208, 235);
+            arrowColours[1, 2] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(238, 244, 252);
+            arrowColours[1, 3] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(229, 237, 247);
+            arrowColours[1, 4] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(223, 234, 247);
+            arrowColours[1, 5] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(241, 246, 254);
+            arrowColours[1, 6] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(243, 247, 252);
+            arrowColours[1, 7] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(250, 252, 255);
 
             // pressed state
-            arrowColours[2, 0] = Color.FromArgb(215, 220, 225);
-            arrowColours[2, 1] = Color.FromArgb(195, 202, 210);
-            arrowColours[2, 2] = Color.FromArgb(242, 244, 245);
-            arrowColours[2, 3] = Color.FromArgb(232, 235, 238);
-            arrowColours[2, 4] = Color.FromArgb(226, 228, 230);
-            arrowColours[2, 5] = Color.FromArgb(230, 233, 236);
-            arrowColours[2, 6] = Color.FromArgb(244, 245, 245);
-            arrowColours[2, 7] = Color.FromArgb(245, 247, 248);
+            arrowColours[2, 0] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Tracking);//Color.FromArgb(215, 220, 225);
+            arrowColours[2, 1] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Tracking); //Color.FromArgb(195, 202, 210);
+            arrowColours[2, 2] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Tracking); //Color.FromArgb(242, 244, 245);
+            arrowColours[2, 3] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Tracking); //Color.FromArgb(232, 235, 238);
+            arrowColours[2, 4] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Tracking); //Color.FromArgb(226, 228, 230);
+            arrowColours[2, 5] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Tracking); //Color.FromArgb(230, 233, 236);
+            arrowColours[2, 6] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Tracking); //Color.FromArgb(244, 245, 245);
+            arrowColours[2, 7] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Tracking); //Color.FromArgb(245, 247, 248);
 
             // background colors
-            backgroundColours[0] = Color.FromArgb(235, 237, 239);
-            backgroundColours[1] = Color.FromArgb(252, 252, 252);
-            backgroundColours[2] = Color.FromArgb(247, 247, 247);
-            backgroundColours[3] = Color.FromArgb(238, 238, 238);
-            backgroundColours[4] = Color.FromArgb(240, 240, 240);
+            backgroundColours[0] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(235, 237, 239);
+            backgroundColours[1] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(252, 252, 252);
+            backgroundColours[2] = _palette.GetBackColor1(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(247, 247, 247);
+            backgroundColours[3] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(238, 238, 238);
+            backgroundColours[4] = _palette.GetBackColor2(PaletteBackStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(240, 240, 240);
 
             // track colors
-            trackColours[0] = Color.FromArgb(204, 204, 204);
-            trackColours[1] = Color.FromArgb(220, 220, 220);
+            trackColours[0] = _palette.ColorTable.StatusStripGradientEnd; //Color.FromArgb(204, 204, 204);
+            trackColours[1] = _palette.ColorTable.StatusStripGradientBegin; //Color.FromArgb(220, 220, 220);
 
             // arrow border colors
-            arrowBorderColours[0] = Color.FromArgb(135, 146, 160);
-            arrowBorderColours[1] = Color.FromArgb(140, 151, 165);
-            arrowBorderColours[2] = Color.FromArgb(128, 139, 153);
-            arrowBorderColours[3] = Color.FromArgb(99, 110, 125);
+            arrowBorderColours[0] = _palette.GetBorderColor1(PaletteBorderStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(135, 146, 160);
+            arrowBorderColours[1] = _palette.GetBorderColor1(PaletteBorderStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(140, 151, 165);
+            arrowBorderColours[2] = _palette.GetBorderColor1(PaletteBorderStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(128, 139, 153);
+            arrowBorderColours[3] = _palette.GetBorderColor1(PaletteBorderStyle.ButtonStandalone, PaletteState.Normal); //Color.FromArgb(99, 110, 125);
 
             //Border colors
             borderColours[0] = _palette.GetBorderColor1(PaletteBorderStyle.InputControlCustom1, PaletteState.Normal);
@@ -392,7 +392,12 @@ namespace Krypton.Toolkit.Suite.Extended.Controls
         /// <param name="state">The <see cref="ScrollBarArrowButtonState"/> of the arrow button.</param>
         /// <param name="arrowUp">true for an up arrow, false otherwise.</param>
         /// <param name="orientation">The <see cref="ScrollBarOrientation"/>.</param>
-        public static void DrawArrowButton(Graphics g, Rectangle rect, ScrollBarArrowButtonState state, bool arrowUp, ScrollBarOrientation orientation)
+        public static void DrawArrowButton(
+           Graphics g,
+           Rectangle rect,
+           ScrollBarArrowButtonState state,
+           bool arrowUp,
+           ScrollBarOrientation orientation)
         {
             if (g == null)
             {
@@ -563,7 +568,10 @@ namespace Krypton.Toolkit.Suite.Extended.Controls
         /// <param name="gripImage">The grip image.</param>
         /// <returns>The adjusted rectangle.</returns>
         /// <remarks>Also rotates the grip image if necessary.</remarks>
-        private static Rectangle AdjustThumbGrip(Rectangle rect, ScrollBarOrientation orientation, Image gripImage)
+        private static Rectangle AdjustThumbGrip(
+           Rectangle rect,
+           ScrollBarOrientation orientation,
+           Image gripImage)
         {
             Rectangle r = rect;
 
@@ -594,7 +602,10 @@ namespace Krypton.Toolkit.Suite.Extended.Controls
         /// <param name="g">The <see cref="Graphics"/> used to paint.</param>
         /// <param name="rect">The rectangle in which to paint.</param>
         /// <param name="state">The <see cref="ScrollBarState"/> of the thumb.</param>
-        private static void DrawThumbVertical(Graphics g, Rectangle rect, ScrollBarState state)
+        private static void DrawThumbVertical(
+           Graphics g,
+           Rectangle rect,
+           ScrollBarState state)
         {
             int index = 0;
 
@@ -806,7 +817,11 @@ namespace Krypton.Toolkit.Suite.Extended.Controls
         /// <param name="rect">The rectangle in which to paint.</param>
         /// <param name="state">The <see cref="ScrollBarArrowButtonState"/> of the arrow button.</param>
         /// <param name="arrowUp">true for an up arrow, false otherwise.</param>
-        private static void DrawArrowButtonVertical(Graphics g, Rectangle rect, ScrollBarArrowButtonState state, bool arrowUp)
+        private static void DrawArrowButtonVertical(
+           Graphics g,
+           Rectangle rect,
+           ScrollBarArrowButtonState state,
+           bool arrowUp)
         {
             using (Image arrowImage = GetArrowDownButtonImage(state))
             {
@@ -826,7 +841,10 @@ namespace Krypton.Toolkit.Suite.Extended.Controls
         /// <param name="rect">The rectangle in which to paint.</param>
         /// <param name="state">The <see cref="ScrollBarArrowButtonState"/> of the arrow button.</param>
         /// <param name="arrowUp">true for an up arrow, false otherwise.</param>
-        private static void DrawArrowButtonHorizontal(Graphics g, Rectangle rect, ScrollBarArrowButtonState state,
+        private static void DrawArrowButtonHorizontal(
+           Graphics g,
+           Rectangle rect,
+           ScrollBarArrowButtonState state,
            bool arrowUp)
         {
             using (Image arrowImage = GetArrowDownButtonImage(state))
@@ -849,7 +867,8 @@ namespace Krypton.Toolkit.Suite.Extended.Controls
         /// </summary>
         /// <param name="state">The button state.</param>
         /// <returns>The arrow down button as <see cref="Image"/>.</returns>
-        private static Image GetArrowDownButtonImage(ScrollBarArrowButtonState state)
+        private static Image GetArrowDownButtonImage(
+           ScrollBarArrowButtonState state)
         {
             Rectangle rect = new Rectangle(0, 0, 15, 17);
             Bitmap bitmap = new Bitmap(15, 17, PixelFormat.Format32bppArgb);
@@ -1007,7 +1026,10 @@ namespace Krypton.Toolkit.Suite.Extended.Controls
         /// <param name="radiusX">The x-radius.</param>
         /// <param name="radiusY">The y-radius.</param>
         /// <returns>A <see cref="GraphicsPath"/> object representing the rounded rectangle.</returns>
-        private static GraphicsPath CreateRoundPath(Rectangle r, float radiusX, float radiusY)
+        private static GraphicsPath CreateRoundPath(
+           Rectangle r,
+           float radiusX,
+           float radiusY)
         {
             // create new graphics path object
             GraphicsPath path = new GraphicsPath();
