+++ Source/Krypton Toolkit/Main/Krypton.Toolkit.Suite.Extended.Scroll.Bars/Krypton.Toolkit.Suite.Extended.Scroll.Bars.licenseheader
@@ -1,0 +1,8 @@
+﻿extensions: .cs
+#region BSD License
+/*
+ * Use of this source code is governed by a BSD-style
+ * license or other governing licenses that can be found in the LICENSE.md file or at
+ * https://raw.githubusercontent.com/Krypton-Suite/Extended-Toolkit/master/LICENSE
+ */
+#endregion
